+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Update datatables to 1.12.1
 - Fix measurement unit always shown ml in amount to make window
 - DEPRECATED: Ingredients card view
+- Add option to print the formula without csv export
 - TECH Preview: PV Online share With a friend
 
 ### Version 4.7